refactor: 멤버 관련 타입 분리
