+++ apps/web/src/constants/memberData.ts
@@ -26,31 +26,7 @@ import {
   메이커스_정효림,
   메이커스_방재현,
 } from "@/components/vision";
-
-export type SupportersRole = "대표" | "운영" | "인프라" | "BX";
-export type MakersRole = "프론트엔드" | "백엔드" | "디자인" | "기획";
-
-export interface SupporterMember {
-  id: number;
-  name: string;
-  role: SupportersRole;
-  imageUrl: string;
-}
-
-export interface MakersMember {
-  id: number;
-  name: string;
-  role: MakersRole;
-  imageUrl: string;
-  description: string;
-}
-
-export interface MemberTabItem {
-  value: string;
-  label: string;
-  gridClassName: string;
-  renderCards: () => React.ReactNode;
-}
+import type { SupportersRole, MakersRole, SupporterMember, MakersMember } from "@/types/ui/member";
 
 export const roleBadgeVariantMap: Record<SupportersRole, ThemeVariant> = {
   대표: "lime",
